Address feedback from reviewers
