+++ spmd/tensor/ops/view_ops.py
@@ -601,6 +601,7 @@ def register_prop_rule_map(
 
 register_prop_rule_map("aten.view.default", Tensor.view)
 register_prop_rule_map("aten.view.SymInt", Tensor.view)
+register_prop_rule_map("aten._unsafe_view.default", Tensor.view)
 register_prop_rule_map("aten.unsqueeze.default", torch.unsqueeze)
 register_prop_rule_map("aten.expand.default", Tensor.expand)
 register_prop_rule_map("aten.permute.default", torch.permute)
